frontend: reduce duplicate code related to run-once capability
Changed RunOnceModel to be abstract and add UserPortalRunOnceModel that
extends it. the code that initialize the run-once-model in the user
portal was moved from UserPortalListModel to it.

Add WebadminRunOnceModel that extends RunOnceModel for the web-admin.
the code that initializes the run-once-model in the web-admin was moved
from VmListModel to it.

Remove duplicate code in the run-once model of the webadmin and
the run-once model of the user-portal, and put the common code in
RunOnceModel instead.

Change-Id: I6561df71f093e2739e689722ead4f85428c04d06
Signed-off-by: Arik Hadas <ahadas@redhat.com>