frontend: add an option of default timezone
This patch adds an empty entry in the timezones list presented in the
UI, which represents the selection of the default timezone that is
defined in the system.

This new empty entry is now selected by default when creating new VM,
and when editing a VM that no timezone was defined for it.

This patch fix bug 922609: before this patch the first timezone was
selected by default when opening the edit dialog of VM with no timezone
defined and therefore when pressing 'ok' the system detected a change in
the timezone field and tried to update it - even when it wasn't possible.
from now on, the empty entry will be selected by default in that case so
no change will be detected unless the user actually changed the timezone
field.

Change-Id: I7f8c80ae783e88740ee907d9ff52c0bd6c40734c
Bug-Url: https://bugzilla.redhat.com/922609
Signed-off-by: Arik Hadas <ahadas@redhat.com>