core: Fix for wrong error massage in add disk
while quota in audit mode, when user tried to add a disk to an SD which
has no assigned quota - an error message was presented and the operation
was blocked. The error message was not correctly formatted.

This happens because when the DC was moved from Enforced to Audit mode
the quota cache was not refreshed. So all quotas on that DC remained in
the old enforcement mode.

Fixed the cache refresh rule. Now that the error is thrown from the
correct place in the code and only in Enforced mode, it is correctly
formatted.

Change-Id: Ifb23898a52c654dd03b7c9d1421fe456c9196da2
Bug-Url: https://bugzilla.redhat.com/906485
Signed-off-by: Ofri Masad <omasad@redhat.com>