userportal: Auto-Connect to RDP on Win 8 and Win 2012-64bit
Connect automatically doesn't take in effect that for Windows 8/2012 RDP is
default.

Steps to Reproduce:
1. Windows 8/2012 VM for user in the UP
2. Connect with auto-connect enabled from the login screen

Spice console appears instead of RDP one.
(As there is performance issue with unaccelerated SPICE console, RDP was agreed
to be default for Windows 8/2012. So make auto-connect to check this
correctly.)

Fixed by utilizing ConsoleUtils and ConsoleManager classes which are used for
connecting to console in regular use case. These classes take into account
that spice is unsupported on the mentioned platforms.

Change-Id: I33d8e4b4701f25bb64a7410512f17c266f872a86
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=918605