engine: Enable gluster hook on cluster
Added Bll command to enable a given gluster hook
on all available up servers in the gluster cluster. This command
will be invoked from the hook management UI.

Once the BLL command is invoked,
- VDS command is called to enable hook on each of the server.
- Hook's status is changed to enabled in DB.
- If there are errors enabling the hook on any of the servers in
the cluster, the conflict status of hook is updated in DB to
indicate that there is a status conflict.

Added an abstract class GlusterHookStatusChangeCommand that
EnableGlusterHookCommand extends from. This class uses
ThreadPoolUtil to invoke the VDS commands simultaneously.

Change-Id: Icc6f9c77393ebed2803ec2a1b295a09f61642c31
Signed-off-by: Selvasundaram <sesubram@redhat.com>
Signed-off-by: Sahina Bose <sabose@redhat.com>