userportal,webadmin: Persist the changes from console popup
Make the changes made in console popup (both WA and UP) persistent
per browser.

The core of the implementation is the ConsoleOptionsFrontendPersisterImpl
which is a gin managed class containing two methods:
- storeToLocalStorage
- loadFromLocalStorage

This methods needs the model to/from to copy the data to/from local
storage and the context, which can be webadmin, userportal's basic
view or userportal's extended view (because the values are configured
separatly for this three places).

The load method copies the selected data from local storage only if
they are still possible in the specific context (for example when the
user selected some spice options, this has been stored to local
storage and than he changed the default console to VNC, the spice
options are not going to be copied).

When using, you need to make sure that you load the stored data
from the local storage before opening the edit console and before
connecting to the console, and also that you store all the data to
local storage after the user changes them in the edit console.

Change-Id: I847e17c6e28db7921f0bc162f0ff70399e671a03
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>