core: remove default locking pairs
This patch remove the default locking pairs in LockMessagesMatchUtil.
The error message in those pairs was too general and not informative
enough, so they should not be used.

The usages of those pairs were replaced with calls to
LockMessagesMatchUtil#makeLockingPair, using the locking group that was
used before and the error message ACTION_TYPE_FAILED_OBJECT_LOCKED, and
the error message ACTION_TYPE_FAILED_OBJECT_LOCKED is now marked as
deprecated.

Change-Id: Ic24c713bb1f38f32a4109b951131b47da20345ea
Signed-off-by: Arik Hadas <ahadas@redhat.com>