core: create 3.0 compatible ovf when export from 3.0 compatible cluster
This patch fix a problem that exporting VM/template from a 3.0
compatibility cluster resulted in a ovf file that wasn't 3.0 compatible,
therefore couldn't be imported in 3.0 environment.

There were 2 changed in the ovf files since 3.0:
1. The disk aliases are no longer in the pattern "Drive 1", "Drive 2"...
2. The usb-policy values changed from ["Enabled", "Disabled"] to
["ENABLED_NATIVE", "ENABLED_LEGACY", "DISABLED"]

So now, the export process checks the compatibility version of the
cluster to which the resource belongs, and create the ovf file with the
values that match that version.

Note that this is a quick-and-dirty fix. it is planned to be
re-implemented with a better design so that the code will be more
maintainable.

Change-Id: I5e7ecfd7c71a5c22ed7751de9f46b690d5236643
Signed-off-by: Arik Hadas <ahadas@redhat.com>