frontend: ConcurrentModificationException in AdElementListModel
The exception occured in the populateRoles method when user is adding
permissions to the VM in permissions subtab. The problem is introduced by
refactoring in 05c06da3 and is only visible in gwt-debug console.

The problem was caused by removing items from the list in the for-each loop.
Fixed by introducing a copy of the modified list (to avoid modifying the
collection passed in the argument).

Change-Id: Id845725507782162c14f6e7b214f65cc6060f59b
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>