engine: fixed NPE in add host
The patch http://gerrit.ovirt.org/13001 modified setVds to set the
mVdsId to Empty in case VDS is passed as null. This results in an
NPE as AddVdsCommand explicitly sets VDS to null, and later calls
getVdsIdRef(), expecting it to be null, but is now Empty.

Hence modifying setVds() to not set it to Empty in case of null, but
leave it as null itself.

Change-Id: I369203472969e707cb515bf593f12849b13b694c
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>