webadmin: Console connect button availability fix
This patch fixes the behavior of console button availability broken by
I00158bbd. For the availability check, the ConsoleUtils is used.
It also removes the event for connecting to console (it was needed because
uicommonweb is not under gin management). Now the UiCommonDefaultTypeResolver
is used for resolving needed classes (adding new interfaces/moving classes was
needed to make this possible).

The signature of the methods of the console options persister is changed - the
console context parameter was removed because it's redundant (can be infered
from the model parameter).

Change-Id: If102303d6b5c5ac45959b87a540c8ea278c227c3
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>