core: Remove VmRunHandler.getPluggedDisks(VM)
Removed VmRunHandler.getPluggedDisks(VM) and replaced it with a direct
DAO call DiskDao.getAllForVm(vmId, true).

This approach has better performance (since all the filtering is done in
the database, and there is no roundtrip via Java) and leaves the code
easier to understand (there's just less of it).

This patch includes:
1. A new method in DiskDAO - getAllForVm(Guid vmId, boolean
   onlyPlugged), its filtered counterpart, the underlying stored
   procedure (naturally), and tests for this method.
2. Removing VmRunHandler.getPluggedDisks(VM)
3. Changing all the usages of VmRunHandler (i.e.,
   CreateAllSnapshotsFromVmCommand and VmRunHandler itself).

Change-Id: I55052dad87a552df200039e111fd0dd6658f6ae8
Signed-off-by: Allon Mureinik <amureini@redhat.com>