core: Fix unpermitted storage related commands in Audit mode
When Data Center is in quota enforcement mode 'Audit', a user can perform
disk actions on a SD which he does not have permission for (e.g. move
disk).

This was caused by wrong permission subjects list creation which
neglected the 'Audit' mode.

Fixed the permission subjects list creation to include 'Audit' mode as
well as 'Disabled' mode (in 'Enforcing' mode this operation will be
blocked by the quota QuotaManager)

Change-Id: Ie50e7dd91611c2b959cea3a78c2684451836760b
Signed-off-by: Ofri Masad <omasad@redhat.com>