packaging: make validateFQDN case insensitive
If the specified FQDN to rhevm-setup is a different case to that
returned by the DNS PTR record, then validateFQDN reports that the
reverse-resolved name doesn't match the supplied name and fails the
validation and hence rhevm-setup fails.

This patch makes validateFQDN case insensitive so it doesn't matter if
the supplied FQDN differs in case to that retured by the DNS PTR record,
because DNS itself is case insensitive.

Change-Id: I7619971956c1eba8b6d3f2a345b8fc85a778384e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=928165
Signed-off-by: Mark Huth <mhuth@redhat.com>