engine: Removing CustomLogFields annotattion (Second super fix in AuditLogDirector)
After some investigation I found out that we don't need CustomLogFields annotaattion at all, this
is a best optimization which can be done here.
We don't need it because of all required names of parameters are located inside our audit log message
and their values will be retrieved via reflection, that's why we don't need to keep additional list of
parameters names inside annotattion.

Benefits:
1. Reduced number of operations for processing annotattion (Improved CPU consaption)
2. Simplified development of audit log messages, no need for annotattion.
3. Code becomes understandable for everyone (at least I hope)

This is a kind of patch that I like:
1. Improved perfromance
2. Code becomes more clear
3. A lot of garbage removed

Change-Id: Ied60d0000e26158baee64c83540d683093485dd8
Signed-off-by: Michael Kublin <mkublin@redhat.com>