engine: Disable gluster hook command
Added bll command to disable gluster hook on all
servers in a cluster.
This command extends from GlusterHookStatusChangeCommand
which does the work of calling the VDS command for each
of the server in the cluster.
The command succeeds if atleast one of the servers returns
success.

Refactored enable and disable hook tests to extend
from a common class.

Change-Id: I6fb3e034013d2ceb3590a61efc9bc46b1c3f378f
Signed-off-by: Selvasundaram <sesubram@redhat.com>
Signed-off-by: Sahina Bose <sabose@redhat.com>