engine,webadmin: Run and pause action enabled
'Start in Pause Mode' checkbox added to the Edit VM properties
dialog, instead of only in the Run Once dialog.

This is needed to assist the admin in boot-level troubleshooting.
Allowing the VM to always start in paused mode without having
to select "Run once" will save the admin time and hassle
when starting and stopping VMs multiple times.

Change-Id: Ib1fe46a9946f7914095310ddac9ab1a9e5a6564b
Signed-off-by: Libor Spevak <lspevak@redhat.com>