core: RemoveImage as sync operation
* When performing RemoveImage command, the engine can treat the vds
command DeleteImageGroup as synchronous operation instead  of asynchronous
(as it does before this patch). The call might be treated as synchronous
because when the call returns, the image is considered to be non
existent on the storage side, we shouldn't wait for the deletion to
end before proceeding with the related db updates.

*When receiving an ImageDoesNotExist error from vdsm during remove image -
proceed with removing the image as it doesn't exist.

*known issue - when performing remove disk, in case of an engine crash
after RemoveImage completed but before the vm's ovf generation was
incremented - the ovf won't be updated.  That's a non issue as the
ovf can contain also today images that where deleted from the storage
and can be fixed in a separate patch to reduce the 'noise' here.

Change-Id: Ic822a92acb8231d2ca84ae092bd98659507925d8
Bug-Url: https://bugzilla.redhat.com/884635
Signed-off-by: Liron Aravot <laravot@redhat.com>