engine: "dual mode" on vm with agent should be reported as "virtio"
vnic with "dual mode" on vm with an agent should be reported to the vdsm as
"virtio" only.
Currently, there is a leftover code that prepares also an interface with
"rtl8139" type. But does nothing with it. So there is not a real bug,
just a code fix.

Also, the vm.getHasAgent() returned false even if there is an agent on the
vm.

The cause of the bug is that the guest-agent-version was updated in
the RunVmCommand.attachCd method only if there was an active ISO domain in
the dc.
In this patch the guest agent version update was moved to the runVm method
to make sure the version is always updated when trying to run the vm.

Change-Id: I072ee66172027d5b4bea1b0ff2f0033d69bf6111
Bug-Url: https://bugzilla.redhat.com/923889
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>