engine : ovirt-node enter installing status when the install fails
When ovirt-node installation fails because of missing iso file
the status of the host is left as Installing.

This patch adds VdsDynamic data to compensation context if can do action
of InstalllVdsCommand fails. The status is reverted back to the previous
status in compensation logic. The error messages are added to the
return values CanDoActionMessages so the message can be displayed
in cli

Change-Id: I2d74e946cb68d2b9dbc11fc5c6bed9f63ba35fe2
Bug-Url: https://bugzilla.redhat.com/947982
Signed-off-by: Ravi Nori <rnori@redhat.com>