packaging: engine-service: do not clean environment
the init system may put important variables in environment.

add LC_ALL as java is effected by AC_CTYPE per defaults of string
conversions, as if it is set by init it will effect us.

in practice we should only set LC_CTYPE or use jnu.encoding, when
we are sure that no default locale is used for conversions within
application.

Change-Id: Ia8231b343574df83e65c55d9ffa75698a3f2d692
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>