packaging: engine-service: clean /var/tmp in upgrade and cleanup
previous service implementation used mixed ownership of root/ovirt in
/var/tmp/ovirt-engine, /var/lib/ovirt-engine/deployments.

the /var/tmp/ovirt-engine was not cleared if daemon exited so likely to
remain.

new service implementation does not run under root account, so having
root owned resources is not healthy.

during upgrade setup correct setting, during setup remove leftovers, as
engine-cleanup not to be trusted as it does not actually revert to
initial state.

Change-Id: If0efa0b8a71673140490e2c363af7d3ba6044ffe
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>