engine: fix NPE when trying to export not existing template
The fix is all about checking whether the template exists before trying
to get its storage pool id.

This patch contains two other changes:
1. the description of the command moved from #canDoAction to the
constructor which is a more appropriate place to initialize things.
2. add a check that the template exists as the first thing that is
checked in #canDoAction(). this check is already made at
MoveOrCopyTemplateCommand#canDoAction but it is taking place too late -
we should address this issue but in a different patch in order to keep
this one simple fix.

Change-Id: I757672ec2f989f4dc1bc5403878c5cbee69f9cf0
Bug-Url: https://bugzilla.redhat.com/928365
Signed-off-by: Arik Hadas <ahadas@redhat.com>