core: fix single sign on for all guests
This patch fix a problem that caused single-sign-on capability not to
work for some guests: vdsm sends the name of the guest agent application
in the applications list. the engine expects to receive RHEV-Agent as
the guest agent application name, but that's true only for guest agent
running on Windows and in d/s version.

This solution is:
1. set all the possible names for the guest agent application, on all
operating systems, in the configuration so that the engine will look for
all of them (instead of only RHEV-Agent).
2. set the relevant names of the guest agent application in u/s, in
ovirt's engine configuration.

Change-Id: If256ddda586eb53af0ca55d6b97dd98e10139d8a
Bug-Url: https://bugzilla.redhat.com/922398
Signed-off-by: Arik Hadas <ahadas@redhat.com>