core: fix overriding placeholder val in audit log
Fix overriding placeholder value set in AugitLogableBase.customValues
with empty string in TypeCompat.
In cases when the value of a placeholder was already populated with
custom value,
no need to invoke a method to try populate it again since in some cases
an empty string returns causing the override of already set non empty value.

Example where bug is shown - storagePoolName is displayed as unknown.
"Host cannot access one of the Storage Domains attached to the Data
Center <UNKNOWN>. Setting Host state to Non-Operational."

Change-Id: I4239085d0bd0246369e93e38a394ab8306c05887
Signed-off-by: Alissa Bonas <abonas@redhat.com>