userportal: redo patch 12748: show the correct VM's console type
The console type presented in VMs lists at the user-portal (basic +
extended) is based on the default display type defined for the VMs,
which is not always right because VM can be run in Run-Once mode
with a display type which is different from the default.

This patch ensure that the right console type is displayed for each VM
in the user portal by changing the way the console type is determined -
if the VM is running then we determine the console type according to the
actual display type (presented by the display type field in VmDynamic),
and according to the default display type, as before, only if the VM is
not running.

Change-Id: I488ceeba48c02e21b19762260a21d0efe1a2eab9
Bug-Url: https://bugzilla.redhat.com/907776
Signed-off-by: Arik Hadas <ahadas@redhat.com>