engine: entity snapshot valiadation
This patch adds a validation to the snapshot mechanism. If there is no
DAO for the entity's class, then the rollback mechanism will not be able
to roll it back, therefore it is be better to fail the transaction
before the compensating transaction tryes and fails to handle the
snapshot.

Change-Id: Ibf28a6d3dd389ab303d2d482e25b94dab4dbafdd
Signed-off-by: Laszlo Hornyak <lhornyak@redhat.com>