core: Update VM status if host is not available
The VM status is set to Unknown if the host is marked as non-responsive
(after timeout). The status icon in the WebAdmin change reflects
the Unknown status. When the host is available again (Up status), the change from Unknown
status to real status is logged into the audit log (VM status restored).

Change-Id: If569de776dfb9c97589603040e7a7f620409c361
Bug-Url: https://bugzilla.redhat.com/921521
Signed-off-by: Libor Spevak <lspevak@redhat.com>