core:we are fencing a host when putting it in...
we are fencing a host when putting it in maintenance after failed reinstall.

There are two commands that handle Host Maintenance

1) MaintenanceVdsCommand
2) MaintenanceNumberOfVdssCommand

In 1) there is handling in canDoAction that prevents the operation for
InstalledFailed status while this handling is omitted from 2)

I would expect that 2) will contain all validation checks done by 1) and
the best will be that they will call the same validation code. but this
is not the case right now.

To keep my change minimal and risk-less, I will have to copy the
relevant validation block from 1) code to 2)

Change-Id: I1ed67cb14b31597d147cee3a64056e3486f51d26
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=906389
Signed-off-by: Eli Mesika <emesika@redhat.com>