core: Run tests without data source location retry
Since commit fd6835 the application retries the location of the data
source, this means that some tests will take a long time to run because
the data source is never deployed in the testing environment. To prevent
that this change adds configuration parameters to run those tests with
an alternative configuration that disables the retry.

Change-Id: I8583da699ae46ce571bb96aa79809031d8d54990
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>