engine: New test method in DBConfigUtilsTest
There seems to be a perception in the community that there is some bug
in the config framework because of which the api returns a null value if
there is no entry in vdc_options for a given config.

This patch attempts to add a new test case that verifies that this is
not the case, and if entry is missing in vdc_options, value returned is
from the @DefaultValueAttribute annotation in ConfigValues enum.

For this purpose,
- Modified DBConfigUtilsTest to extend from BaseDAOTestCase, so that
  fixtures.xml can be used to insert entries in vdc_options
- Modified fixtures.xml to insert two entries in vdc_options for the
  config NonVmNetworkSupported for version 3.0 and 3.2
- Renamed DBConfigUtils#RefreshVdcOptionCache to refreshVdcOptionCache
  and made it protected so that it can be called from test class passing
  the test dbFacade
- Added new test method testGetValue() to verify that default value is
  returned if entry is not present in vdc_options

Change-Id: Ia736419e766bc6ce2e82158f775817e885b90dfd
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>