webadmin: Gateway is not defined after bonding the RHEVM interface.
When detaching the network its parameters (boot protocol, address, subnet
and gateway- for managenemt network) are saved for the next attach.

The bug was that in case of managemnet network the gateway parameter was
always determined from the nicEntity.
But if the management network is over vlan it should be taken from the
vlanNic.

Change-Id: I7deff4305d04a6cfc943bb38cb795373ea7ff6e4
Bug-Url: https://bugzilla.redhat.com/923443
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>