engine: Improving memory footprint - JSON part
The following patch should imprtove a used memory in all work related to JSON.
The scenarious which will be improved: all scenarios which are using compensation and vm monitoring.
Motivation is due to wiki of Jackson serializer it is thread safe, except configuration part,
but we are not changing configuration (Also I did short code review in order to verify that ObjectMapper is thread safe)
ObjectMapper is not lightweight object and also has some internal cache in order to improve it performance, so
a change is obvious: serializer and desirializer which are based on ObjectMapper should be single instance.

Change-Id: I11a76b46a5bfa269605249c3dfc2d69a55d8bbef
Signed-off-by: Michael Kublin <mkublin@redhat.com>