frontend: reduce duplicate code related to run-once capability - cont
This patch is the last part in the effort to reduce the duplicate code
related to run-once capability. The creation of UICommands that are
related to run-once capability was moved into the run-once models, and
the event-handling code that is tightly related to the run-once model
was also moved into the run-once models.

Change-Id: I774524ca7bd4af35d8c3140e2ce23f369d727e28
Signed-off-by: Arik Hadas <ahadas@redhat.com>