engine: AssignQuota button under disks refresh bug
In case a quota mode set to !Disabled (or vice-versa) the
Assign Quota button (in various disk tabs) availability wasn't refreshed.

The problem caused because we relied on the selected DC tree
item quota mode, which not getting refreshed in current ui infrastructure.
Instead we will fetch quota mode from disk BE itself.

A bug was fixed in DB to reference data center even
if there is no quota for the disk (using the storageDomain DC instead
of quota table).

To allow new functionality for all tabs, introduced a new interface
and static public utility method to avoid code duplication.

Change-Id: I21a0886cfe82d441ad93fd9b15d72f8a4e63002c
Bug-Url: https://bugzilla.redhat.com/950519
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>