userportal,webadmin: disable SPICE proxy has no effect
Edit console dialog -> enable SPICE proxy -> connect to VM->
edit console dialog again -> disable SPICE proxy -> connect to VM

After this, the SPICE will connect to the VM using the proxy,
because it remembered it from the previous setting.

Fixed by resetting the Proxy property if the proxy is not enabled.

Change-Id: I5931b4242bdba7755e29c3942a1df2eab03f7b5f
Bug-Url: https://bugzilla.redhat.com/924605
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>