userportal: always update console from local storage
The details view of the basic view shows the SPICE instead of RDP
even the RDP has been selected in the edit console.

The problem was, that the UserPortalItemModel has been updated from
local storage only:
- before the popup dialog has been opened
- before the console connect has been called

This patch adds this update right after the UserPortalItemModel is
created so it will be up to date right after loading the data from
the server.

Change-Id: I87e0cd9d8b1912845b5b7ae5ebedf3a9b02bd2b5
Bug-Url: https://bugzilla.redhat.com/924610
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>