engine: Add non-plugin RDP invocation
This patch adds the possibility to run RDP console without MS RDP
activeX plugin installed (can be used e.g. by Firefox in MS Windows).
For this the attachment servlet is used which generates RDP descriptors.

The feature is configured via console popup dialog. Default (per engine)
configuration is stored in vdc_options.

NOTE: The automatic login feature will not work with non-plugin
invocation, because there is no straightforward way to pass a password
in the RDP descriptor.

NOTE on the db upgrade.
Since the ClientConsoleModeDefault was related to SPICE only, this patch
renames this option to ConsoleModeSpiceDefault and adds similar option
for RDP (ClientModeRdpDefault).

Change-Id: Ibbfc1b143b7125c278d22326785476d71d53ec0d
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=918890