core: Fix the enable-dao-tests
Change Ia736419e added a database test to DbConfigUtilsTest, which was
not a database test originally.
This causes the regular test suite to fail since no database is
present.

The following patch amends the dal module's pom not to run this test by
default, but only when -Penable-dao-tests is specified.
This is done by adding this test to the default ignore list, and
overriding the surefire-plugin configuration in the profile.

Change-Id: Icbc55a014a54c584a476e66612e54b8b926d02c4
Signed-off-by: Allon Mureinik <amureini@redhat.com>