2. core : Distingushing between entity and vdsm task ids
This patch is first in series of patches to handle the required
mechanism for better support for cases of missing tasks upon
engine crash.
One of the steps in the proposed solution is to diffrentiate between
an entity ID provided by engine and the vdsm task ID provided by
VDSM.
The first patch simply adds the vdsmTaskId column, and since older code relied
on this id (even though inserted to a "task_id" column) this patch
suggested appropriate fixes.
One of the issues that will be handled in future patches is not to include
the VDSM task ID at the insert_AsyncTasks stored procedure, but at this point
this is not addressed (as this will be a part of a patch handling splitting the
insert of task from getting vdsm task id + updating in db)

http://www.ovirt.org/Features/AsyncTaskManagerChanges_3.3
Change-Id: I75e35773b18e342dbe61242829ce989b1f51ee70
Signed-off-by: Ravi Nori <rnori@redhat.com>