core: Unfilter ImagesHandler.PerformImagesChecks
Before this patch, ImagesHandler.PerformImagesChecks received a
collection of Disks, and filtered them so only DiskImages that aren't
sharable were evaluated.

This patch removes the filtering from this methods, and makes it the
responsibility of the calling method to perform any required filtering.
This patch also adds said filtering wherever relevant, so no behavior is
changed.

This achieves two goals:
1. Added flexibility, to allow possible validation on sharable disks.
2. A clearer API - it is now clear, in compile time, that this method
   only performs validations on DiskImages, as opposed to the way it
   was before the patch, where the function could be passed a faulty
   (e.g., non-existent) LunDisk, which would just be silently ignored.

Change-Id: Icc93d1bb72daed053f0f37ae3c3f12efef915669
Signed-off-by: Allon Mureinik <amureini@redhat.com>