core: Remove checkIsValid from PerformImagesChecks
The recent changes to ImagesHandler.PerformImagesChecks have left the
checkIsValid parameter completely useless:
If it's true, both checkImagesIllegalInVdsm and checkImagesExist are
evaluated, and their respective tests are run if they are true.
If it's false, neither test is run.
In conclusion - it has no extra meaning, and can be removed.

This patch removes it from PerformImagesChecks' signature and all its
usages.
In case both other booleans are false, the function call is removed
altogether.
In case removing this function allows to remove some additional
parameters from the calling function (e.g., as is the can in
VmRunHandler), that is done as well.

Change-Id: I0055151c619dcf9e62144d280714c165a01408e8
Signed-off-by: Allon Mureinik <amureini@redhat.com>