core: PerformImagesChecks no check for no disks
Removed the check for no image disks from PerfomImagesChecks: creating a
template, creating a snapshot and removing a snapshot should also work
on diskless VMs, however useless it may be.

Move VM has this check explicitly defined in its canDoAction, and is
deprecated, regardless.

After this change, there is no difference between PerformImagesChecks
and checkDiskImages, so the ltater (which was private) was removed,
and the code inlined into PerformImagesChecks.

Change-Id: Ib4638c8dc5ba6990cdf0b335c76645ee9dedbeee
Signed-off-by: Allon Mureinik <amureini@redhat.com>