core: block VM migration during live storage migration
This is a best-effort fix for preventing a VM from being migrated during
migration of one of its disks (live storage migration), and vice versa.

One way is already covered - live storage migration is made only when
the VM state is up or paused so it can't be executed when the VM is in
migration.

This patch is partial fix for the opposite way - a check is added to
MigrateVmCommand to verify that none of the VM's disks is locked, which
is the case when one of its disks is being migrated (or during snapshot
creation for the disk).

This fix doesn't cover a situation where the VM migration and storage
migration/snapshot creation are taking place simultanously, and not
cover the case where the VM migration fails thus being rerun and the VM
state is switched to 'UP' in the middle of the process.

Change-Id: Id630a01368290ba9ac6c71c5cab2095f84fa018b
Bug-Url: https://bugzilla.redhat.com/878131
Signed-off-by: Arik Hadas <ahadas@redhat.com>