userportal: prestarted VMs not updated
Have a pool with prestarted VMs. Take a VM from it in user portal.

The result is that the title of the VM gets changed, but nothing else
making it impossible to connect to it.

The problem was, that the PoolItemBehavior.PostTakeVm() loaded
the taken VM from the server and replaced the entity in the
UserPortalItemModel. This does not trigger the
DataBoundTabModelProvider.updateData. On the next refresh cycle,
the optimization logic detects no change in this entity (because
it has really been replaced) and does not trigger the re-rendering
of it.

As the logic of loading the just taken VM from the server was only
in this one place making the behavior incosistent comparing to
all other actions like start/stop/pause etc this logic has been
removed leaving the re-rendering to the refresh cycle.

Change-Id: Ia910ac50675d4a1e8db0f4fd01ce50b93a8d9f34
Bug-Url: https://bugzilla.redhat.com/953690
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>