core: removal of unneeded member/method
StorageDomainCommandBase has a private member called _storagePool and a method
called getStoragePool() which overrides the getStoragePool() method
(which has the same implementation) defined at AuditLogableBase. The setter
method isn't being overriden, which means that calling getStoragePool() on
StorageDomainCommandBase and lower classes in it's hierarchy will return the
value of the _storagePool member defined in StorageDomainCommandBase, while
using the setter will set the value of the storage_pool member defined in
AuditLogableBase, thus it won't affect the value of the recieved storage
pool.

In this patch I just removed the storage_pool member defined within
StorageDomainCommandBase as it's unneeded (we already have that higher
in the hierarchy) and it's getter as they are unneeded.

Change-Id: Ie4746c35e2f5c37e672e47e56314d115c6b25025
Signed-off-by: Liron Aravot <laravot@redhat.com>