core: Allow creating ISO domain on other file-based storages except NFS
This patch allows creating ISO domain on file-based storage domain,
including localfs and posixfs. Even though localfs can't be shared
among the hosts in cluster, it could be useful in the case of no NFS available.
VM can be created on the host which has the ISO domain attached,
and then be migrated to any other host in the cluster or make a template
from it. This is also a good feature for POCs, all-in-one, etc.

Change-Id: I2a8d3ea8ab4ac10353ec8574287458e8eb63e882
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Signed-off-by: Allon Mureinik <amureini@redhat.com>