engine: fetch gluster host uuid during InitVdsOnUp
When a gluster host comes up, if gluster server uuid of the host is empty
it will be fetched and stored in the database. And the host will be
moved to NonOperational if the gluster server uuid is not found in db as well
as failed to retrieve the same from vdsm.

A new NonOperationReason is added and also the corresponding entries to AuditLog as well.

Change-Id: Ied8e8b006adc5722cd70ee19d75a4b6f783d5f44
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>