set correct storage domain id
When executing RecoveryStoragePoolCommand, the master storage domain is
loaded within the CDA method for the various checks - later on, if we do
get for executing the master domain id might have changed meanwhile.
The reload is done also today, but outside of the event - in this patch
i just moved it to happen within till further changes to that
mechanism - as it prevented the added domain to be marked the master,
version not be incremented..etc.

Change-Id: I7cce29e172dc7923efdd50997abe586b506f2fbf
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=955569
Signed-off-by: Liron Aravot <laravot@redhat.com>