webadmin: Making network related subtabs resizeable
Fixed sub tabs-
- Network Cluster
- Network Host
- Network Vm
- Network Template
- Netwrok Permissions (there is some general code used by all
permissions sub-tabs, it will be fixed in a separate patch)
- Vm Interface (The left pane is fixed, the right pane is using
EntityModelCellTable which doesn't have support for resizing).
- Template interface
- Cluster Interface
- Dc Interface

Not Fixed sub tab-
Host Interface- this is unusual table constructed of multiple panel that
look together as a table. The is no infrastructure to support resizing
of such kind of table.

Change-Id: I04c6db892934fa94060bc126b2b317cb651fff32
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>