userportal: closing RDP crashes IE
The problem was, that the MsRdpClient object has been removed from the body in
the "OnDisconnected" event of the MsRdpClient which is too soon and causes the
IE to crash - possibly because IE starts to destroy this object but after
OnDisconnected callback returns, IE tries to give the control back to the
MsRdpClient.

Solved by just removing this line becaue the proper cleanup is already done in
connect() method which removes the previous MsRdpClient if there is one.

Change-Id: I6501bbd88ed77df2d3f8acb7fe61913a2888ce7f
Bug-Url: https://bugzilla.redhat.com/909109
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>