engine: Allow engine to configure management network
The patch eliminates* the need of ovirt-host-deploy to
create the management network on top of the installed
host.

The management network will be created as the last step
of the installation process, by the logical network
definition in the data-center.

Since the network creation relies on the
setupNetworks, no further action will be taken for
non-supported cluster versions.

* ovirt-nodes are the exceptional in this process and for
them (regardless their version), the ovirt-host-deploy
will create the management bridge (regardless the logical
network definition).

If the network is defined otherwise, the admin will have
to use the 'Sync Network' option via the setupNetworks
dialog to enable it.

Once ovirt-node will not create the br* bridges, the engine
will be able to use the same logic for configuring the
management network on the node.

Change-Id: Iaf82e10481e595d690e7ce894283b4ed5b9b3269
Bug-Url: https://bugzilla.redhat.com/891056
Signed-off-by: Moti Asayag <masayag@redhat.com>